requirements.txt
asgiref==3.8.1
astroid==3.3.5
blinker==1.4
certifi==2020.6.20
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.9
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.4
django-filter==24.3
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==1.16.0
filelock==3.16.1
httplib2==0.20.2
importlib-metadata==4.6.4
isort==5.13.2
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.7
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pbr==6.1.0
pillow==11.0.0
pipenv==11.9.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pymongo==3.11.0
pyparsing==2.4.7
pypi==2.1
python-apt==2.4.0+ubuntu4
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
stevedore==5.3.0
systemd-python==234
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.27.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wadllib==1.3.6
zipp==1.0.0
python-3.11.0
gunicorn==21.2.0